+++ src/Yesod/Paginator.hs
@@ -1,111 +1,64 @@
-{-# LANGUAGE CPP #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE TypeFamilies #-}
--------------------------------------------------------------------------------
 -- |
 --
--- Inspiration from a concept by ajdunlap:
---      <http://hackage.haskell.org/package/yesod-paginate>
+-- There are two pagination functions. One for arbitrary items where you provide
+-- the list of things to be paginated:
 --
--- But uses an entirely different approach.
+-- @
+-- getSomeRoute = do
+--     -- 10 items per page
+--     pages <- 'paginate' 10 =<< getAllThings
 --
--- There are two pagination functions. One for arbitrary items where you
--- provide the list of things to be paginated:
+--     defaultLayout $ do
+--         [whamlet|
+--             $forall thing <- 'pageItems' $ 'pagesCurrent' pages
+--                 ^{showThing thing}
 --
--- > getSomeRoute = do
--- >     things' <- getAllThings
--- >
--- >     (things, widget) <- paginate 10 things'
--- >
--- >     defaultLayout $ do
--- >         [whamlet|
--- >             $forall thing <- things
--- >                 ^{showThing thing}
--- >
--- >             <div .pagination>
--- >                  ^{widget}
--- >             |]
+--             $# display at most 5 page elements, with current page middle-ish
+--             ^{'simple' 5 pages}
+--             |]
+-- @
 --
--- And another for paginating directly out of the database, you provide
--- the same filters as you would to @selectList@.
+-- And another for paginating directly out of the database, you provide the same
+-- arguments as you would for @'selectList'@:
 --
--- > getSomeRoute something = do
--- >     -- note: things is [Entity val] just like selectList returns
--- >     (things, widget) <- runDB $ selectPaginated 10 [SomeThing ==. something] []
--- >
--- >     defaultLayout $ do
--- >         [whamlet|
--- >             $forall thing <- things
--- >                 ^{showThing $ entityVal thing}
--- >
--- >             <div .pagination>
--- >                  ^{widget}
--- >             |]
+-- @
+-- getSomeRoute something = do
+--     pages <- runDB $ 'selectPaginated' 10 [SomeThing ==. something] []
 --
--- Both functions return a tuple: the first element being the list of
--- items (or Entities) to display on this page and the second being a
--- widget showing the pagination navagation links.
+--     defaultLayout $ do
+--         [whamlet|
+--             $forall thing <- 'pageItems' $ 'pagesCurrent' pages
+--                 ^{showThing $ entityVal thing}
 --
--------------------------------------------------------------------------------
+--             ^{'simple' 5 pages}
+--             |]
+-- @
+--
 module Yesod.Paginator
-    ( paginate
-    , paginateWith
+    (
+    -- * Type-safe numerics
+      PageNumber
+    , PerPage
+    , ItemsCount
+
+    -- * Paginated data
+    , Pages
+    , pagesCurrent
+
+    -- * The current page
+    , Page
+    , pageItems
+
+    -- * Paginators
+    , paginate
     , selectPaginated
-    , selectPaginatedWith
-    , module Yesod.Paginator.Widget
-    ) where
 
-import Yesod
-import Yesod.Paginator.Widget
+    -- * Widgets
+    , simple
+    , ellipsed
+    )
+where
 
-paginate :: Yesod m => Int -> [a] -> HandlerT m IO ([a], WidgetT m IO ())
-paginate = paginateWith defaultWidget
-
-paginateWith :: Yesod m
-             => PageWidget m
-             -> Int
-             -> [a]
-             -> HandlerT m IO ([a], WidgetT m IO ())
-paginateWith widget per items = do
-    p <- getCurrentPage
-
-    let tot = length items
-    let  xs = take per $ drop ((p - 1) * per) items
-
-    return (xs, widget p per tot)
-
-selectPaginated :: ( PersistEntity val
-#if MIN_VERSION_persistent(2, 5, 0)
-                   , PersistEntityBackend val ~ BaseBackend (YesodPersistBackend m)
-#else
-                   , PersistEntityBackend val ~ YesodPersistBackend m
-#endif
-                   , PersistQuery (YesodPersistBackend m)
-                   , Yesod m
-                   )
-                => Int
-                -> [Filter val]
-                -> [SelectOpt val]
-                -> YesodDB m ([Entity val], WidgetT m IO ())
-selectPaginated = selectPaginatedWith defaultWidget
-
-selectPaginatedWith :: ( PersistEntity val
-#if MIN_VERSION_persistent(2, 5, 0)
-                       , PersistEntityBackend val ~ BaseBackend (YesodPersistBackend m)
-#else
-                       , PersistEntityBackend val ~ YesodPersistBackend m
-#endif
-                       , PersistQuery (YesodPersistBackend m)
-                       , Yesod m
-                       )
-                    => PageWidget m
-                    -> Int
-                    -> [Filter val]
-                    -> [SelectOpt val]
-                    -> YesodDB m ([Entity val], WidgetT m IO ())
-selectPaginatedWith widget per filters selectOpts = do
-    p   <- lift getCurrentPage
-    tot <- count filters
-    xs  <- selectList filters (selectOpts ++ [OffsetBy ((p-1)*per), LimitTo per])
-
-    return (xs, widget p per tot)
+import Yesod.Paginator.Pages
+import Yesod.Paginator.Paginate
+import Yesod.Paginator.Widgets